instructpix2pix and dreambooth ckpting
